refactor: add conditions for user level and authentication

+++ backend/app/Http/Controllers/ProductController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\ProductRequest;
 use App\Models\Product;
+use App\Models\User;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
@@ -12,9 +13,11 @@ class ProductController extends Controller
 {
 
     private $product;
+    private $loggedUser;
 
     public function __construct(Product $product) {
         $this->product = $product;
+        $this->loggedUser = auth()->user();
     }
 
     /**
@@ -31,7 +34,7 @@ class ProductController extends Controller
             ->with('user')
             ->paginate((int) $request->per_page);
 
-            return response()->json($product, Response::HTTP_OK);
+        return response()->json($product, Response::HTTP_OK);
     }
 
     /**
@@ -51,8 +54,11 @@ class ProductController extends Controller
     public function show(String $id): JsonResponse
     {
         $product = $this->product->findOrFail($id);
+        if($this->loggedUser->id === $product->user_id || $this->loggedUser->role === User::USER_ADMIN) {
+            return response()->json(['message' => 'Produto encontrado', 'product' => $product], Response::HTTP_OK);
+        }
 
-        return response()->json(['message' => 'Produto encontrado', 'product' => $product], Response::HTTP_OK);
+        return response()->json(['error' => 'Acesso não autorizado'], Response::HTTP_FORBIDDEN);
     }
 
     /**
@@ -62,9 +68,13 @@ class ProductController extends Controller
     {
         $data = $request->validated();
         $product = $this->product->findOrFail($id);
-        $product->update($data);
 
-        return response()->json(['message' => 'Produto atualizado com sucesso', 'product' => $product], Response::HTTP_OK);
+        if($this->loggedUser->id === $product->user_id || $this->loggedUser->role === User::USER_ADMIN) {
+            $product->update($data);
+            return response()->json(['message' => 'Produto atualizado com sucesso', 'product' => $product], Response::HTTP_OK);
+        }
+
+        return response()->json(['error' => 'Acesso não autorizado'], Response::HTTP_FORBIDDEN);
     }
 
     /**
@@ -73,8 +83,12 @@ class ProductController extends Controller
     public function destroy(String $id): JsonResponse
     {
         $product = $this->product->findOrFail($id);
-        $product->delete();
 
-        return response()->json(['message' => 'Produto removido com sucesso', 'product' => $product], Response::HTTP_OK);
+        if($this->loggedUser->id === $product->user_id || $this->loggedUser->role === User::USER_ADMIN) {
+            $product->delete();
+            return response()->json(['message' => 'Produto removido com sucesso', 'product' => $product], Response::HTTP_OK);
+        }
+
+        return response()->json(['error' => 'Acesso não autorizado'], Response::HTTP_FORBIDDEN);
     }
 }
